#DELETE-PAGE feat: Confirm dialog box

+++ src/main/gui/Windows/DeletePage.java
@@ -3,6 +3,8 @@ package main.gui.Windows;
 import javax.swing.*;
 
 import app.use_cases.DeleteLibrary;
+import app.use_cases.GetLibrary;
+import core.entities.Library;
 import infrastructure.repositories.jackson.LibraryRepository;
 
 import java.awt.Dimension;
@@ -12,24 +14,28 @@ import java.awt.Font;
 import java.awt.event.*;
 
 public class DeletePage implements ActionListener {
+    private final GetLibrary REPOSITORY_FIND;
     public final DeleteLibrary REPOSITORY;
 
-    JFrame frame = new JFrame();
-    JLabel title = new JLabel("Deletar dados");
-    JLabel subtitle = new JLabel("Preencha os dados do valor que deseja excluir: ");
+    private JFrame frame = new JFrame();
+    private JLabel title = new JLabel("Deletar dados");
+    private JLabel subtitle = new JLabel("Preencha os dados do valor que deseja excluir: ");
 
-    JLabel labelCep = new JLabel("CEP: ");
-    JLabel labelNumber = new JLabel("Número:");
+    private JLabel labelCep = new JLabel("CEP: ");
+    private JLabel labelNumber = new JLabel("Número:");
 
-    JTextField txtCep = new JTextField();
-    JTextField txtNumber = new JTextField();
+    private JTextField txtCep = new JTextField();
+    private JTextField txtNumber = new JTextField();
 
-    JButton btnFinish = new JButton("Excluir");
-    JButton goBack;
+    private JButton btnFinish = new JButton("Excluir");
+    private JButton goBack;
 
-    JLabel status = new JLabel("");
+    private int confirmDialog;
+
+    private JLabel status = new JLabel("");
 
     public DeletePage(LibraryRepository repository) {
+        this.REPOSITORY_FIND = new GetLibrary(repository, repository);
         this.REPOSITORY = new DeleteLibrary(repository);
 
         this.title.setBounds(10, 5, 350, 30);
@@ -46,10 +52,10 @@ public class DeletePage implements ActionListener {
         this.labelNumber.setFont(new Font("Serif", Font.BOLD, 18));
         this.txtNumber.setBounds(140, 140, 160, 25);
 
-        this.status.setBounds(30, 200, 350, 25);
-        this.status.setFont(new Font("Serif", Font.PLAIN, 18));
+        this.status.setBounds(30, 200, 450, 25);
+        this.status.setFont(new Font("Serif", Font.PLAIN, 16));
 
-        this.btnFinish.setBounds(250, 220, 100, 40);
+        this.btnFinish.setBounds(250, 240, 100, 40);
         this.btnFinish.setFocusable(false);
         this.btnFinish.addActionListener(this);
 
@@ -63,7 +69,7 @@ public class DeletePage implements ActionListener {
         this.frame.add(this.status);
         this.frame.add(this.btnFinish);
 
-        this.frame.setSize(400, 350);
+        this.frame.setSize(420, 350);
         this.frame.setLayout(null);
         this.frame.setVisible(true);
 
@@ -79,10 +85,33 @@ public class DeletePage implements ActionListener {
     public void actionPerformed(ActionEvent e) {
         if (e.getSource() == this.btnFinish) {
             try {
-                REPOSITORY.deleteLibrary(this.txtCep.getText(), 
-                                        Integer.parseInt(this.txtNumber.getText())
-                );
-                this.status.setText("Respositório deletado com sucesso.");
+                if (this.txtCep.getText().isEmpty() || this.txtNumber.getText().isEmpty())
+                this.status.setText("Ambos os campos CEP e Número devem ser preenchidos!");
+                else {
+                    Library library = REPOSITORY_FIND.getLibraryBy(this.txtCep.getText(), Integer.parseInt(this.txtNumber.getText()));
+                    
+                    confirmDialog = JOptionPane.showOptionDialog(frame, "" + library.getName() +
+                                                                 ", " + library.getEmail() + 
+                                                                 ", " + library.getComplement() + 
+                                                                 ", " + library.getCep() + 
+                                                                 ", " + library.getNumber(), 
+                                                                 "Deseja mesmo excluir?", 
+                                                                 JOptionPane.OK_CANCEL_OPTION, 
+                                                                 JOptionPane.QUESTION_MESSAGE, 
+                                                                 null, 
+                                                                 null, 
+                                                                 null);
+                    if (confirmDialog == JOptionPane.CANCEL_OPTION)
+                        return;
+                    else {
+                        REPOSITORY.deleteLibrary(this.txtCep.getText(), 
+                                                Integer.parseInt(this.txtNumber.getText())
+                        );
+                        this.status.setText("Respositório deletado com sucesso.");
+
+                    }
+
+                }
             } 
             catch(Exception error) {
                 this.status.setText(error.getMessage());
